amo_get/requirements.txt
blinker==1.9.0
click
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
requests==2.32.5
Werkzeug==3.1.3
fastapi
uvicorn
pathlib
aiohttp